requirements.txt
alembic==1.7.7
anyio==3.6.2
apache-airflow==2.2.5
apache-airflow-providers-ftp==2.1.0
apache-airflow-providers-http==2.1.0
apache-airflow-providers-imap==2.2.1
apache-airflow-providers-sqlite==2.1.1
apispec==3.3.2
argcomplete==2.1.2
async-generator==1.10
attrs==20.3.0
Babel==2.11.0
blinker==1.5
cached-property==1.5.2
cachelib==0.6.0
cattrs==1.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.0.4
clickclick==20.10.2
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
connexion==2.14.2
contextvars==2.4
croniter==1.4.1
cryptography==40.0.2
dataclasses==0.8
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.4
dnspython==2.2.1
docutils==0.16
email-validator==1.3.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.19.1
gunicorn==21.2.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.1.0
itsdangerous==1.1.0
Jinja2==3.0.3
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
packaging==21.3
pendulum==2.1.2
pep562==1.1
prison==0.2.1
psutil==5.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.1.0
pyrsistent==0.18.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
requests==2.27.1
rfc3986==1.5.0
rich==12.6.0
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.41.1
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.2.2
termcolor==1.1.0
text-unidecode==1.3
typing==3.7.4.3
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.16
Werkzeug==1.0.1
wrapt==1.15.0
WTForms==2.3.3
zipp==3.6.0
